Reffactoeing making slider an individual component
